Split batch worker shared execution helpers
